Moved 99% map setup into MapsViewModel
